Factorize particle reading for boxlib datasets
