+++ yt/frontends/boxlib/fields.py
@@ -4,6 +4,7 @@ import string
 import numpy as np
 
 from yt.fields.field_info_container import FieldInfoContainer
+from yt.frontends.boxlib.misc import BoxlibSetupParticleFieldsMixin
 from yt.units import YTQuantity
 from yt.utilities.physical_constants import amu_cgs, boltzmann_constant_cgs, c
 
@@ -186,7 +187,7 @@ class NyxFieldInfo(FieldInfoContainer):
     )
 
 
-class BoxlibFieldInfo(FieldInfoContainer):
+class BoxlibFieldInfo(FieldInfoContainer, BoxlibSetupParticleFieldsMixin):
     known_other_fields = (
         ("density", (rho_units, ["density"], None)),
         ("eden", (eden_units, ["total_energy_density"], None)),
@@ -224,26 +225,6 @@ class BoxlibFieldInfo(FieldInfoContainer):
         # "burnstate",
         # "luminosity",
     )
-
-    def setup_particle_fields(self, ptype):
-        def _get_vel(axis):
-            def velocity(field, data):
-                return (
-                    data[(ptype, f"particle_momentum_{axis}")]
-                    / data[(ptype, "particle_mass")]
-                )
-
-            return velocity
-
-        for ax in "xyz":
-            self.add_field(
-                (ptype, f"particle_velocity_{ax}"),
-                sampling_type="particle",
-                function=_get_vel(ax),
-                units="code_length/code_time",
-            )
-
-        super().setup_particle_fields(ptype)
 
     def setup_fluid_fields(self):
         unit_system = self.ds.unit_system
