+++ yt/frontends/chombo/fields.py
@@ -6,6 +6,7 @@ from yt.fields.field_info_container import (
     particle_vector_functions,
     standard_particle_fields,
 )
+from yt.frontends.boxlib.misc import BoxlibSetupParticleFieldsMixin
 from yt.units.unit_object import Unit
 from yt.utilities.exceptions import YTFieldNotFound
 
@@ -25,7 +26,7 @@ class ChomboFieldInfo(FieldInfoContainer):
 # Orion 2 Fields
 # We duplicate everything here from Boxlib, because we want to be able to
 # subclass it and that can be somewhat tricky.
-class Orion2FieldInfo(ChomboFieldInfo):
+class Orion2FieldInfo(ChomboFieldInfo, BoxlibSetupParticleFieldsMixin):
     known_other_fields = (
         ("density", (rho_units, ["density"], None)),
         ("energy-density", (eden_units, ["total_energy_density"], None)),
@@ -63,26 +64,6 @@ class Orion2FieldInfo(ChomboFieldInfo):
         ("particle_luminosity", ("", [], None)),
         ("particle_id", ("", ["particle_index"], None)),
     )
-
-    def setup_particle_fields(self, ptype):
-        def _get_vel(axis):
-            def velocity(field, data):
-                return (
-                    data[(ptype, f"particle_momentum_{axis}")]
-                    / data[(ptype, "particle_mass")]
-                )
-
-            return velocity
-
-        for ax in "xyz":
-            self.add_field(
-                (ptype, f"particle_velocity_{ax}"),
-                sampling_type="particle",
-                function=_get_vel(ax),
-                units="code_length/code_time",
-            )
-
-        super().setup_particle_fields(ptype)
 
     def setup_fluid_fields(self):
         from yt.fields.magnetic_field import setup_magnetic_field_aliases
